Analyse the O.D. values (regression)

Performs a 4-PL regression of the standard values and converts the O.D. into concentration values.

od_to_concentration(id, value, std_key = "^STD", dec = ".")

Arguments

id
a character vector containing the identifiers.

value
a numerical vector containing the O.D. values.

std_key
a character string specifying the common starting pattern of standard point ids (default = "^STD").

dec
a character string used as a decimal separator for the encoded standard concentration values.

Value

A numerical vector containing the calculated concentrations.

Details

A complete example on how to perform an analysis can be found at https://github.com/koncina/elisar.

Examples

# NOT RUN {
library(tidyverse)
library(elisar)

read_plate("od_measure.xls") %>%
  mutate(concentration = get_concentration(id, value))
# }