Import the OD/luminescence/fluorescence measures from plate reader exported excel sheets (like Fluostar Optima, or Tecan Sunrise)

Reads excel files exported from the MARS software (http://www.bmglabtech.com/en/products/software/mars-data-analysis/) or Tecan Magellan (http://lifesciences.tecan.com/products/software/magellan_data_analysis_software) and modified to include the plate layout (see documentation).

read_plate(path, na = "")

Arguments

path
vector containing the path(s) to the input file(s)

na
character vector of strings to use for missing values (see readxl::read_excel)

Details

Example on how to prepare the excel file can be found at https://github.com/koncina/elisar.

Examples

# NOT RUN {
library(elisar)

# Import file(s)
df <- read_plate("od_measure.xls")
df <- read_plate(c("od_measure1.xls", "od_measure2.xls"))
# }